Fixed vol container group and onprem test problems
The members of a container group were stepping on each other in the
volume storage temporary file.  The temporary directory is now
group/container.  Clean up is protected via a lock.

The onprem tests do not have a clean initial state so the tests need
to be more selective about verifying that a file exists and deleted.